# Optimise WAL loading by removing extra map and caching min-time