--- a/tsdb/head_test.go
+++ b/tsdb/head_test.go
@@ -239,6 +239,8 @@ func BenchmarkLoadWAL(b *testing.B) {
 						require.NoError(b, err)
 						h.Init(0)
 					}
+					b.StopTimer()
+					w.Close()
 				})
 		}
 	}