@bboreham 2021-08-05
It's failing a WAL corruption unit test now, which I shall attempt to debug through, but if the answer is instantly apparent to someone with more familiarity in tsdb please shout.