@bboreham 2021-08-07
The failure was due to skipping `mint`/`maxt` in `processWALSamples()`.
I changed things around so the head's min and max times reflect all mmapped chunks added up to that point.
This likely does more atomic operations, but I think that is justified by the flow being easier to follow.

Incidentally I question why `memSeries.maxTime()` does not look at mmapped chunks.
Also why `memSeries.minTime()` returns `MinInt64` as a fallback - elsewhere the fallback for min time is `MaxInt64`.