@codesome 2021-08-10
> I question why memSeries.maxTime() does not look at mmapped chunks.

There will always be a head chunk with at least 1 sample (we m-map a chunk when we are going to add a sample to the head chunk that will be newly created). But that is only when things are running normal, if there was WAL corruption and no head chunk exists, then we should actually look at m-map chunks too :) so, something to fix here.

> Also why memSeries.minTime() returns MinInt64 as a fallback - elsewhere the fallback for min time is MaxInt64.

Not sure, need to check the usage